materials/RON_MON/pyproject.toml
[project]
name = "ron-mon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "catboost>=1.2.10",
    "hydra-core>=1.3.2",
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "mlflow>=2.22.0",
    "optuna>=4.8.0",
    "pandas>=3.0.2",
    "plotly>=6.7.0",
    "polars>=1.39.3",
    "rdkit>=2026.3.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
]
